TFP-4340 Tar inn om behandling krever sammenhengende uttak og startet på ny delregel
